requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
attrs==19.3.0             # via pytest
coverage==5.0.4           # via pytest-cov
ddt==1.3.1                # via -r requirements/test.in
importlib-metadata==1.6.0  # via pluggy, pytest
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pathlib2==2.3.5           # via pytest
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-cov==2.8.1         # via -r requirements/test.in
pytest-django==3.9.0      # via -r requirements/test.in
pytest==5.4.1             # via -r requirements/test.in, pytest-cov, pytest-django
pytz==2019.3              # via -r requirements/base.txt, django
six==1.14.0               # via packaging, pathlib2
sqlparse==0.3.1           # via -r requirements/base.txt, django
wcwidth==0.1.9            # via pytest
zipp==1.2.0               # via importlib-metadata
